BUG#34038210 MySQL-Shell core dumped when use rejoinInstance
In a scenario where an instance server_uuid is changed, trying to rejoin
that instance to its previous ReplicaSet results in a core dump.

This core dump is due to a missing check, where an uninitialised
topology member is used. This patch fixes this by making sure that
rejoin updates the metadata with the up-to-date uuid of the instance to
rejoin.

Change-Id: I62fec258e93c32c9fce8fd679dbce05f8c90ed43