Bug#33941759 Metadata upgrade ignores prompt "No" to abort upgrade
When upgrading from version 2.0 and if routers are present in the
metadata, when asked if the user wishes to proceed with the upgrade,
the answer is ignored and the upgrade proceeds regardless.

The patch fixes this behaviour so that, if the user answers "No", the
upgrade is indeed aborted.

Additionally, the question asked was tweaked to that the possible
answers aren't duplicated: "... (Yes/No) [y/N]"

Change-Id: I5d63ae23e75097a396e739b08e001dec18849903