BUG#32618416 ANOTHER SHELL VALGRIND WARNING DURING ADDINSTANCE/CLONE
Fixed crash caused by usage of unitialized value:

A sscanf() used in the code to get the cursor position in the screen
used a char array as buffer that wasn't initialized. Depending on the
implementation of sccanf() this can cause a crash.

This patch fixes that by memsetting the array with null terminating
characters.

Change-Id: I5cfcd75389621d7a1e993328c63ac52f027190b2