BUG#33239404 Incomplete cleanup after dissolve on cluster that was in a clusterset
Using Cluster.dissolve() on a Replica Cluster that has been removed from
a ClusterSet using the force option (because it was unreachable or the
replication channel was not found or stopped) did not ensure all
recovery and ClusterSet replication accounts were dropped and the
Metadata schema was cleared.

Considering that a Cluster.dissolve() can only be used in the following
scenarios:

  - Standalone Cluster
  - Replica Cluster forcedly removed from ClusterSet

... the command must ensure that all recovery and ClusterSet replication
accounts are removed and the Metadata schema is removed too. This patch
extends the command's behavior to do that instead of just clearing the
Cluster and its instances records from the Metadata and dropping just
the recovery accounts assoaciated with those instances.

Added regression tests to cover both scenarios.

Change-Id: I706c4a297129d89d1ffd52c967e581978fce761d