BUG#34002328 BUG#31693765 replicationLag related bugs
BUG#31693765 MYSQL SHELL REPORTED REPLICATIONLAG IS NOT CORRECT

The replicationLag value reported in Cluster.status() was not being
calculated properly resulting in several issues:

  - When a large transaction is being applied the calculated value is
  wrong and does not increase as the lag increases
  - The value obtained does not order workers by newest applied
  transaction returning the first one obtained
  - When the applier is idle, either the last "known" lag is reported or
  "NULL"

This patch fixes the issues by introducing a new query that addresses
those issues ensuring that:

  - If the replication connection or the SQL thread is not running,
  returns 'null'
  - When the last queued transaction and the last applied transaction
  are the same, or when the applying transaction is 0 it reports that
  the applier queue has applied everything: 'applier_queue_applied'
  - Calculates the lag as being the difference between NOW() and the
  time the transaction being applied was commited by its source
  - When multiple workers are being used, obtains the information from
  the worker that is executing the oldest transaction to report its lag

BUG#34002328 status() replicationLag 'null' is confusing

replicationLag reported 'null' in most scenarios on which the applier
was idle which is very dubious as it can lead to multiple
interpretations.

This patch fixes as well as described above.

Change-Id: Ic25e3d16bb9d2cdee1ace07eca40c50b8779b1eb