BUG#33795149 BUG#33795166 Destructive error when unset attribute in JSON using CRUD operations
When using X DevAPI CollectionModify object, methods which accept a
document path as a parameter did not properly handle arguments which
contained spaces, i.e.:
  col.modify("_id='1'").unset("some path")
would attempt to remove an attribute named "some".

The input arguments of these methods were not validated correctly, paths
with spaces need to be specified using a path expression, i.e.:
  col.modify("_id='1'").unset("$.'some path'")
or need to be quoted using backtick characters:
  col.modify("_id='1'").unset("`some path`")

With this patch, paths containing spaces which are not properly
specified using a path expression (i.e. "some path", "'some path'") are
treated as errors.

Change-Id: I0a5b46f9b873823dca2b4cc02121ff236082f05a