Fixed uncommon crash the clone monitoring code:
monitor_post_clone_gr_recovery_status() uses a pointer returned from
monitor_clone_recovery() that can possibly be null when a restart
happens while executing the code.

This patch fixes that by ensuring the returned pointer is never null,
i.e. if a restart already happened and the instance came back use a
pointer to a session to the instance otherwise use the initial pointer
with the current still open session.

Change-Id: I140171e9950f2fc1dce282bd3303bafde2daf2df