WL#14862 BUG#33381060 BUG#30884646
AdminAPI - Guarantee API objects freshness

Throughout the AdminAPI, there are validations and operations that may
be executed on an outdated view of the topology through its object
(Cluster/ReplicaSet/ClusterSet) resulting in either unexpected failures
or erroneous results.

To work around that, users would have to obtain a "fresher" object
before performing the operations, i.e.:

  - dba.getCluster()
  - dba.getReplicaSet()
  - dba.getClusterSet()

This patch changes that by improving the whole usability of the AdminAPI
to ensure that validations, updates, and operations on objects already
instantiated are performed on the most up-to-date view of the topology.

Considering that every AdminAPI operation must perform a precondition
check, this patch introduces the topology detection and update during
that precondition check.

This patch also resolves the problem of the excessive number of sessions
and metadata refreshes happening with each command executed. Several
refactors of the usage of Connection pools were done.

Updated existing tests to remove the workarounds of obtaining "fresh"
objects and added new tests.

Change-Id: I7beb25c167a655dad81e6bae3d3a15a9ecb2ac00