Fix failing test scenario in cluster_status.js
In MySQL 8.0.31 (BUG#34231291) it's now forbidden to start/stop replica
for the Group Replication applier channel.

The test scenario was validating that 'replicationLag' is null when the
applier channel is stopped by stopping it manually. That is now
forbidden, and for the channel to be reported as being OFF GR must be
either stopped or in ERROR state. 'replicationLag' is not displayed in
those cases so the tests must be executed only if the Server version is
< 8.0.31.

Change-Id: I99b3c5be6770f80971f5c3c0812cff1bb6ff09bd