Bug #34321108 Failure loading files with CRLF line ending in JavaScript mode
Loading of text files for JavaScript operations with files with window's
style line endings (CRLF). The failue is caused because a validation is done to ensure
the number of read characters matches the file size, however, line endings are
interpreted as a single character, resulting in load error.

This patch fixes the problem with 2 actions: reading the files in binary format and
validating based on read error bit flag verification rather than counting read characters.

Change-Id: I5907392c440de82ae8bca282fbd7b5f10769428b