Bug #34353095 Removed support for global initialization scripts in Windows.
The Shell is able to load initialization scripts from a global and a user context, this is compatible with the standard MySQL deployments in Linux/Macos but not with the standard MySQL deployment in Windows.

This patch removes the usage of global initialization scripts in Windows.

Whenever a global initialization script is detected in Windows, a WARNING will be shown to indicate the user that the initialization script will be ignored.

Change-Id: Ia2fe42890d92272c032fa107d73b2755000260b1