BUG#34370211 BUG#34342946 SQL code completion suggests candidates for the previous token
BUG#34370211 use mysql <tab> -> mysql , ... -> invalid SQL

Pressing a tab:
  use mysql <TAB>
would suggest schema names, which is not correct in this case.

BUG#34342946 SQL code completion suggests columns when it shouldn't

Pressing a tab:
  SELECT mysql.plugin.name <TAB>
would suggest 'dl' and 'name' column names, which is not correct in this
case.

The root cause of both bugs was the same: code completion was suggesting
rules as candidates, even though the whole rule was already consumed. As
a fix, rule which has already consumed the whole available input is not
taken into account.

Change-Id: I6924cd24a293fcef6f28f8de1808f767202975e7