BUG#34403275 Accessing a non existing key in a python dictionary results in extra error
Accessing a non existing key in a Shell dictionary using the get()
method resulted in an additional SystemError exception:

shell.parse_uri("root:@localhost").get('y')
IndexError: invalid key: y

The above exception was the direct cause of the following exception:

Traceback (most recent call last):
  File "<string>", line 1, in <module>
SystemError: <built-in method get of shell.Dict object at 0x7fa458731c90> returned a result with an error set

The implementation of get() method was not correct, it reported a Python
exception and then returned a value, which is not expected by Python.
Additionaly, the get() method should not throw if the requested key does
not exist, but return None.

Change-Id: I17e048ad543635ceed23f0fc7d858c60fbbeefa3