BUG#34408687 Dba.rebootClusterFromCompleteOutage: Could not find metadata for Cluster
When rebooting a Cluster that belongs to a ClusterSet, where the
topology of the ClusterSet changed since the Cluster was last online
(i.e.: another Replica Cluster was added to the ClusterSet), results in
an exception.
This was caused by attempting to obtain the information of the
Clusters added to the ClusterSet, in the one that's being rebooted,
where that information isn't available.

This patch fixes this scenario by properly dealing with the exception.

Change-Id: I05dadc967fddfdeb781186dedd4a851b27cc2e09