BUG#34408669 BUG#34409792 Loading a dump fails if user does not have privileges to set sql_generate_invisi
When loading a dump into a 8.0.30+ instance, shell is using the
sql_generate_invisible_primary_key session variable in order to control
creation of primary keys (but only if the global value is different from
the requested one). If user who is loading the dump doesn't have the
privileges required to set sql_generate_invisible_primary_key session
variable, the load operation will fail. As a fix, if global value is
OFF and the primary keys should be created but user does not have
required privileges, shell will fall back to pre 8.0.30 behaviour (and
edit the CREATE TABLE statement and add a primary key). If the global
value is ON, and the primary keys should NOT be created, but user does
not have the required privileges, the behaviour is not changed, unless
the MYSQLSH_ALLOW_ALWAYS_GIPK environment variable is set.

Change-Id: I5086b1453da4de494534976a7a0e1b134f972abe