BUG #34385244 Review instance verification check in rebootClusterFromCompleteOutage
With WL#14862, the way rebootClusterFromCompleteOutage was retrieving
info regarding instances of the Cluster and of its ClusterSet (if it
belonged to one) can be simplified, leading to leaner code and less API
calls.

This patch includes the following changes:
- use instance type to validate the cluster instances, instead of
creating and retrieving its cluster metadata
- removed code that looped, trying to reconnect to the target instance,
to check if it was invalidated
- simplified the code that rejoins the Cluster back into its ClusterSet
- added a new test to check that the 'force' option is valid when the
instance is an INVALIDATED replica.

Change-Id: Ifdccbbb97eb3e8f2c544fcd4efbf05a53c693e23