BUG#33931291 dba.upgradeMetadata refers to InnoDB Cluster when upgrading a ReplicaSet
This patch updates the message output from dba.upgradeMetadata to be
more generic, so that explicit references to either "Cluster" or
"ReplicaSet" are no longer present.

This patch also includes a fix to dba.dropMetadataSchema() to prevent
duplicated messages from being printed, when the instance has GR
installed but is disabled.

Change-Id: I601063721ea1f886d2135dd1f4a25e6a09da7880