BUG#34049624 Detect invalid CREATE USER syntax
Dumping users in MariaDB and loading the dump into MySQL results in
errors, because MariaDB uses some incompatible CREATE USER syntax. In
order to prevent this, if MariaDB is detected and 'users' dumpInstance()
option is set to true, an error is reported, stating that this option
needs to be set to false.

A new dumper error code has been added to handle this case:

  "Dumping user accounts is currently not supported in MariaDB. Set the " \
  "'users' option to false to continue."

Change-Id: Ia68e821872751a0716139c5f6643c29cee43d559