BUG#34324165 Incorrect globalStatus reported in cluster.status()
When the primary Cluster in a ClusterSet is in a state other than OK,
the "globalStatus" output in ClusterSet.status() for a replica is
NOT_OK, even if the replica is in a valid state. This is incorrect
because the global state of the replica does not depend on the global
state of the primary.

This patch fixes this issue, which means that, regardless of the global
state of the primary, the replica will return its global state,
including OK_NOT_REPLICATING if it's OK, just not replicating because
the primary is offline.

Other changes:
 - fixed how primary is checked in check_preconditions, which lead to
some updates in the preconditions of some APIs.
 - removed static declarations from preconditions.h

Change-Id: I79e17e92d9a9e27de33924524c9550e11d615693