BUG#34264094 cluster.rejoinInstance() fails When ClusterSet is under Heavy Load
Rejoining an instance back to a Replica Cluster that is using the
'MySQL' communication stack and the Primary Cluster of the ClusterSet is
under a significant load results in a failure and the instance is unable
to rejoin the Cluster.

The problem relies on the fact that during the rejoin process the
replication account must be re-created and that happens at the Primary
Cluster. However, if the Primary Cluster is under load that transaction
might take some time to be replicated to the Replica Cluster and it
might not be there yet at the time the instance attempts to join back
the replicaton group. When using the 'MySQL' communication stack Group
Replication needs to do pre-authentication using that recovery account
before actually attempting to join the group so that will fail since
the account wasn't replicated yet.

This patches fixes that by ensuring that, when using the 'MySQL'
communication stack, the Cluster's transaction-set is synchronized with
the Primary Cluster's one after the replication account is re-created.

Added regression tests.

Change-Id: Ic6dc4ad153201a02da0c8584372b615a2c1dab25