BUG#34063868 Shell ERROR 1101 (42000): BLOB, TEXT, GEOMETRY or JSON can't have a default
If a dump created from a 5.7 instance contained a table with a BLOB,
TEXT, GEOMETRY or JSON column with a default value, it would fail to
load.

In order to mitigate this issue, upgrade checker now checks if such
tables exist and reports an error if a table is found. Upgrade checker
is automatically executed when dumping with the 'ocimds' option set to
true, this is the only case when dumper/loader can detect this issue.
Users can also execute the upgrade checker manually, which is suggested
if they plan to migrate from a 5.7 instance to 8.0 one.

Change-Id: I1d34410b59406c2c068d01c1e957857a77b9dfed