BUG#34446932 Unnecessary restart of clusterset replication channel after primary switchover
During the switchover of the primary member of a Cluster, if the Cluster
belongs to a ClusterSet and is a Replica Cluster, the AdminAPI does the
following:

1. Stops the ClusterSet replication channel, otherwise, GR wouldn't
allow the switch since there are running replication channels in the
instance

2. Calls the UDF for the switchover: group_replication_set_as_primary()

3. Starts the ClusterSet replication channel again

However, step 3 is unnecessary since GR will automatically restart the
channel in such a scenario since it's configured to be a managed
replication channel.

Currently, the operation is a no-op. However, there's a remote
possibility that it causes undefined behavior in some particular scenarios.

This patch removes the unnecessary step.

Change-Id: I2e9e4d52e543b817e2f39446edcb607d09fca510