BUG#34190956 BUG#34458017
BUG#34190956 Support stats_updates_frequency option in ClusterSet.setRoutingOptions()

In 8.0.31, MySQL Router was extended to allow configuring the frequency
of periodic (last check-in) updates that it does in the metadata when
the target Cluster belongs to a ClusterSet (WL#15230).

Considering that this is a Metadata option, the AdminAPI must support
setting and visualizing it. For that reason, this patch extends
<ClusterSet>.setRoutingOption() to allow setting it and
<ClusterSet>.routingOptions([router]) to display the information about
it.

BUG#34458017 setRoutingOption to null resets all options

<ClusterSet>.setRoutingOption(option, null) resets all routing options
to the default value instead of just the target option.

This patch fixes that by ensuring only the target option is reset to the
default value and all remaining options are kept unchanged.

Added regressions tests.

Change-Id: I6c6134e18b27f437098e3b31f0d513cb398c17f3