BUG#33799352 I want to change the schema and load a dump using util.dumpSchemas.
Previously, it was only possible to use the 'schema' option to load a
dump created by util.dumpTables() function. With this change, it is
possible to use this option with any dump and load it into a new schema
(effectively renaming the source schema), as long as only one schema is
loaded (either the dump contains just one schema, or filtering options
result in just one schema).

Warning: the loaded SQL is not parsed in any way, and the old schema
name is not replaced with the new schema name. This means that if i.e. a
view or routine or any other DB object references the old schema name,
it is going to reference the same name after the dump is loaded.

Note: when using any of the load filtering options, the old schema name
should be used.

Note: if the new schema does not exist, it is going to be created.

Change-Id: Iac13e7e395cf1de1a6ae0d4f65cc018a5c7b439b