BUG#34424385 Allow to use Shell to set ipWhitelist
Since version 8.0.24, MySQL server allows the system variable
group_replication_ip_allowlist to be changed while GR is running.

This patch adds support to set the option "ipAllowlist" either in the
cluster (which updates all instances) or just in a single instance.

Also included in this patch is a fix for Cluster.setOption, where the
instances weren't being checked if they were online and if the option
specified was supported on each of them.

Change-Id: Id4f9d8837abd4678547d0bf7b385123592e2c122