BUG#34465132 clusterset.status() Attempt to read null value
If a Replica Cluster is completely offline, but also the Group
Replication plugin was uninstalled from all of its members,
<Clusterset>.status() fails with a logic error.

The problem was caused by attempts to query Group Replication sysvars
that are not available when the plugin is not installed, resulting in
internal exceptions.

This patch fixes that by ensuring that those queries are safe regardless
if the plugin is installed or not, i.e. when the plugin is not installed
the variable gets assigned to a default empty string.

Added regression tests.

Change-Id: I10a8790d560e553b6a70e0ee8b42e4dac4d080dc