Bug #34465882	forcePrimaryCluster failing, no override
Failover was failing on high load because the code to check whether the
cluster being promoted is the most up-to-date was comparing the promoted
cluster with itself. The comparison with itself was failing because the
applier was still catching up, resulting in different GTID_EXECUTED
values when it's queried at 2 different times.

Fix is to skip the promoted cluster itself when checking for the most
up-to-date cluster.

Change-Id: I4a13246537f36b2cd57db55c3aca80da4719c496