BUG#34460776 Segmentation fault -- cluster rescan {--addInstances:'auto'}
While processing the command line arguments, and in case of an error,
the shell would try and print to the log file. But at this point of the
shell initialisation, the log file hasn't been created yet, resulting in
a crash.

This patch fixes this problem by preventing the log from being used by
the command line parser (i.e.: before it is initialised).

Change-Id: I3763e65c1acbc6fb814afbf02af61efafe22f766