BUG#34457903 Not possible to change cluster options in a clusterset
Setting tags in a Cluster that belongs to a ClusterSet, via
<Cluster>.setOption() or <Cluster>.setInstanceOption() fails with an
error regarding the Cluster status in the ClusterSet.

This was caused by a bug in the preconditions checker handling for those
commands. When setting global options, all members of the target Cluster
must be online, however, when setting tags it's OK that not all members
are online as long as there is quorum.

For that reason, the preconditions check for those commands use a custom
Function_availablility struct. When ClusterSet was implemented, and the
Cluster_global_status_mask was added to the struct, this custom struct
wasn't updated.

This patch fixes that and adds regression tests.

Change-Id: Ib9ac243d255f657ce2007cbbb3c6a1ca50ef267c