BUG#34465006 createReplicaCluster() must reset member actions to avoid upgrade scenario bug
When a Cluster is upgraded from 8.0.26 to 8.0.27+, every instance that
is taken out and rejoined back will have the list of available Group
Replication member actions reduced to the ones available in 8.0.26
(the current primary). In this particular scenario, the action
'mysql_start_failover_channels_if_primary' is removed since it's only
available in 8.0.27+.

As soon as the primary is taken out, upgraded, and rejoined, the member
actions available in the group are the "reduced" ones, meaning
'mysql_start_failover_channels_if_primary' is not available.

Attempting to create a Replica Cluster in that instance results in an
error since that member action is configured and enabled in Replica
Clusters.

This is a bug in Group Replication: BUG#34464526. Group Replication
should reset the member actions to the available ones according to the
version.

To avoid the bug, createReplicaCluster() must reset the member actions
in the target instance by making use of the UDF
'group_replication_reset_member_actions()'.

Change-Id: I80bcf281faf796d06f71f85ec709811df17ae867