BUG#34417802 ClusterSet: fenceWrites() fails in a split brain scenario
On a split-brain scenario due to a network partition, fencing the old
primary when Shell can access both clusters fails with an error
indicating that using fenceWrites() on REPLICA Clusters is forbidden.

The full scenario is:

1) Network split
2) Forced failover to Cluster B;
  2.1) Cluster A is invalidated
  2.2) A is a primary cluster from its point of view
3) MySQL Shell @ A can access both clusters and, because of WL#14862
gets the metadata info from current primary: B
  3.1) B is the primary Cluster, A is a Replica invalidated
4) A.fenceWrites() -> error: "operation not permitted on REPLICA
Clusters"

This bug was unmasked with WL#14862 that ensures that operations are
performed always on the most up-to-date view of the topology.

This patch fixes the bug by allowing the usage of cluster.fenceWrites()
on REPLICA Clusters too. There's no reason why the operation should be
blocked on REPLICA Clusters. If the Cluster is an healthy REPLICA then
the operation is a no-op. If any of its members has super_read_only
disabled, the operation is useful to set it back.

On the other hand, it must be forbidden to unfence from writes any
REPLICA Cluster so this patch adds that validation to
cluster.unfenceWrites().

Added regression tests.

Change-Id: I30b5f1a71b2e0236b43a88c4efefb09cfda8b676