BUG#34173880 When load volume is very high, the throughput of a stand-alone load is worse than parallel load with dump
When loading a dump which contained a large number of big tables, load
performance of sequential process (dump first, load when finished) was
significantly slower than the performance of a concurrent load (load
while dump is ongoing).

This problem was caused by a bug in the scheduling algorithm, it always
tried to pick a table to be scheduled which is currently not being
loaded. If number of tables exceeded the number of available worker
threads, this resulted in frequent changes of active tables, which in
turn caused frequent index reloads and had negative impact on the InnoDB
buffer pools.

As a fix, loader tries to load only data into N tables at a time, where
N is set to the number of worker threads (the `threads` load option).

After this fix, the load time for the case described in the bug report
(100 tables, each with 5 million rows) has decreased from 1h 10min +/-
10% to 32min 16sec +/- 1sec.

Change-Id: I398197714b27415dedec1b7dfd17ae5dbc74429d