BUG#34479702 Unable to load files larger than 4GB
When loading a dump/importing a table from a remote storage on Windows,
if the file size exceeds 4GB, the operation fails with an error:
    stoul argument out of range

The problem was caused by the code which reads the Content-Length
header, it was using the std::stoul() function which returns an unsigned
int. Since this type on Windows is 4 bytes long, this caused an
exception whenever the size of loaded file exceeded 4GB. The code has
been updated to use std:stoull().

Change-Id: I227a7cb60b2e0e65b6f9722798346347a067d034