Fix usage of exit in child process
After a call to fork(), if the process is the child process, one should
use _exit() instead of exit() to prevent possible concurrency with the
parent process (e.g.: double flushing of stdio buffers, etc.).

This patch replaces calls to exit() with _exit() to prevent these
scenarios. Also tweaked the log file descriptor to avoid being used by a
child process when it is exec().

Change-Id: Ic3df96452c1b0a6bfbd0ed9852c1dbc276bce9b5