Bug#34525709 Unable to extend the utils global object with utilities using plugins
This patch resolves the problem of not being able to extend the global utils object.

The utils object has been turned into an extensible object to allow adding functions
and extension objects through shell plugins.

In addition, the @plugin decorator got a new optional "parent" parameter to allow
the user to specify an existing extension object as parent for the new plugin object
in case it is not desired as a global object.

Change-Id: Ibe71ec3cb0d8c8339d2053f8fb49ccb48fded4a5