Bug#31460929	PYTHON MULTIPROCESSING NOT WORKING
Fixed -c option to pass cmdline arguments to the executed command.

The subprocess module forks itself using the -c cmdline argument, which
will lead to the entry point of the multiprocess module in child
processes. The previously added -c shell option was missing forwarding
of additional command line arguments to the command.

Change-Id: I792e0e12fb858ca3bd7fca2e826154c5ede92aba