WL#15311 BUG#34606598 BUG#32669773 where clause option for exportTable()
New options added to the `util.exportTable()` method
====================================================

* `where` - string (default: not set) - A valid SQL condition expression
  used to filter the data being exported.
* `partitions` - list of strings (default: not set) - A list of valid
  partition names used to limit the data export to just the specified
  partitions.

New options added to the `util.dump*()` methods
===============================================

* `where` - dictionary (default: not set) - A key-value pair of a table
  name and a valid SQL condition expression used to filter the data
  being exported.
* `partitions` - dictionary (default: not set) - A key-value pair of a
  table name and a list of valid partition names used to limit the data
  export to just the specified partitions.

The table name specified by the key should be a valid MySQL identifier
in the format of `schema.table`, quoted using backtick characters when
required.

If a key specifies a table which is not included in the dump or does not
exist, or a table which belongs to a schema which is not included in the
dump or does not exist, it is be ignored.

If a table included in the dump is not listed in either `where` or
`partitions` options, the whole data from that table is going to be
dumped.

Options copied to the `util.dump*()` methds from `util.exportTable()`
====================================================================

* `fieldsTerminatedBy`: string (default: `\t`) - This option has the same
  meaning as the corresponding clause for SELECT ... INTO OUTFILE.
* `fieldsEnclosedBy`: char (default: `''`) - This option has the same
  meaning as the corresponding clause for SELECT ... INTO OUTFILE.
* `fieldsEscapedBy`: char (default: `\`) - This option has the same
  meaning as the corresponding clause for SELECT ... INTO OUTFILE.
* `fieldsOptionallyEnclosed`: bool (default: false) - Set to true if the
  input values are not necessarily enclosed within quotation marks
  specified by `fieldsEnclosedBy` option. Set to false if all fields are
  quoted by character specified by `fieldsEnclosedBy` option.
* `linesTerminatedBy`: string (default: `\n`) - This option has the same
  meaning as the corresponding clause for SELECT ... INTO OUTFILE.
* `dialect`: enum (default: `default`) - Setup fields and lines options
  that matches specific data file format. Can be used as base dialect
  and customized with `fieldsTerminatedBy`, `fieldsEnclosedBy`,
  `fieldsEscapedBy`, `fieldsOptionallyEnclosed` and `linesTerminatedBy`
  options. Must be one of the following values: `default`, `csv`, `tsv`
  or `csv-unix`.

The data files have the `.tsv` file extension if `dialect` is set to
`default` or `tsv`, the `.csv` extension if `dialect` is set to `csv` or
`csv-unix`, and the `.txt` extension in all the remaining cases.

Change-Id: I4e6c9532c37f2b5d1f7e2629b381012195e60327