BUG#34579287 dba.getCluster() fails if primary is unreachable
When the primary member of a Cluster fails so it's unreachable but
failover hasn't happened yet, dba.getCluster() fails with an error
indicating no primary member has been found.

The command should not fail but instead, attempt to obtain the Cluster
object using a secondary instance.

This patch fixes that by ensuring the command automatically retries
using a secondary member when:

  - Group Replication still reports having a Primary member even though
  the instance is unreachable
  - Group Replication does not report having a Primary member because
  automatic failover is still not complete

Added regression tests.

Change-Id: I283961836f12244d0c06b68a27eaa107508292e3