BUG#34584939 Adding instance to a new replica cluster under load results in errors
Attempting to add an instance to a freshly created Replica Cluster, when
the Primary Cluster is under high load, fails with several errors
related to super_read_only.

The issue happens when a new Cluster handle is obtained right after a
Replica Cluster is created, and the addInstance() operations is called
using that handle.

The root cause of issue is that .createReplicaCluster() does not sync
the Metadata update transactions to ensure those were already propagated
to the new Replica Cluster. Any attempt to obtain a Cluster handle may
result in the object having a wrong view of the topology and be
considered a Standalone Cluster, i.e. the Metadata server is pointing to
the Cluster's primary member instead of the ClusterSet's primary member.

This patch fixes that by ensuring the missing transaction sync is done.

Added regressions tests.

Change-Id: I45053042dbb60e34d702daf34f30ed3fb4e377db