Refactoring URI parsing/encoding.
This patch introduces a couple of interfaces for the serialization operations
on data defined in a URI:

- IUri_parsable: to process the data contained on a URI into a parsable object.
- IUri_encodable: to form an URI based on the data contained in an encodable object.

The intention of this patch is to provide a clean way to integrate the URI
parsing/encoding services for the current objects:

- Connection_options
- Ssh_connection_options

Adds support for parsing generic and file URIs through the following objects:

- Generic_uri
- File_uri

And provides a clean interface for future objects requiring parsing, encoding or both.

Additionally, several bugs regarding proper encoding/decoding of URI components were fixed.

Change-Id: I55db70ce8a4dba0bc6653f377924cc3493b5fff7