WL#15337 Support for Azure Blob backend for Import/Export
New options
===========

The new options have been added to the following methods:

* `util.exportTable()`
* `util.importTable()`
* `util.dumpInstance()`
* `util.dumpSchemas()`
* `util.dumpTables()`
* `util.loadDump()`

`azureContainerName`
--------------------

Use specified Azure Container for the location of the dump.

If the option is used, the dump is stored in the specified Azure Container.
The connection is established using the azure settings which can be
specified through any of:

- The options listed here.
- The followinh environment variables:

  - `AZURE_STORAGE_ACCOUNT`
  - `AZURE_STORAGE_KEY`
  - `AZURE_STORAGE_CONNECTION_STRING`
  - `AZURE_STORAGE_SAS_TOKEN`

- The options defined in the configuration file.

The directory structure is simulated within the object name.

`azureConfigFile`
-------------------

By default, the standard configuration file located at `~/.azure/config`
are used, but a custom configuration file can be specified with this option.

1. The following setting are read from the `storage` of the configuration file:

  * `sas_token`
  * `connection_string`
  * `account`
  * `key`

`azureStorageAccount`
---------------------

Defines the Azure Storage Account to be used in the operation

`azureStorageSasToken`
---------------------

Define a temporary Shared Access Storage token to sign the request used in the
operation.

Connection Configuration Resolution
===================================

It is a requirement to sign the requests done on the operation, this could be
achieved by using a SAS Token or using a Shared Access Key.

The configuration to be used to sign the requests will be resolved in the
following order of precedence:

  *  SAS Token
  *  Connection String
  *  Individual Settings (Account/Key)

The SAS Token will be resolved in the following order of precedenve:

  * The `azureStorageSasToken option.
  * The `AZURE_STORAGE_SAS_TOKEN` environment variable.
  * The `sas_token` option in the configuration file.

The Connection String will be resolved in the following order of precedence:

  * The `AZURE_STORAGE_CONNECTION_STRING` environment variable.
  * The `connection_string` option in the configuration file.

The Account will be resolved in the following order of precedence:

  * The `azureStorageAccount` option.
  * The `AZURE_STORAGE_CONNECTION_STRING` environment variable.
  * The `connection_string` option in the configuration file.
  * The `AZURE_STORAGE_ACCOUNT` environment variable.
  * The `account` option in the configuration file.

The Key will be resolved in the following order of precedence:

  * The `AZURE_STORAGE_CONNECTION_STRING` environment variable.
  * The `connection_string` option in the configuration file.
  * The `AZURE_STORAGE_KEY` environment variable.
  * The `key` option in the configuration file.

Change-Id: I6462701e82183a78dce5fba7106a5f3b4995288a