BUG#34395705 Improve error message for metadata node address inconsistency
In 8.0.29, a fix was introduced to prevent invalid endpoints ("grLocal")
from being stored in the metadata. The fix, however, didn't take into
account possible invalid values already being present in the metadata,
hence the reported behavior.

This patch provides a solution for this scenario by displaying a warning
in cluster.status() output for the target instance, and fixing the issue
when cluster.rescan() is executed. Any other command executed after
rescan() won't fail because of an invalid "grLocal" value in the
metadata.

Change-Id: I173c0f5f4f3e63dbb29936bed118e0eedca26dfe