BUG#34465675 finding outdated most recent replica during reboot or failover
In any given instance where the current load is high, there's a change
that the applier lags behind, meaning that GTID_EXECUTED might not
have the most up-to-date GTID set. This means that some APIs, like
ClusterSet.forcePrimaryCluster(), when checking which instance has the
most up-to-date GTID_EXECUTED, need to wait for the applier to finish.

This patch adds an "timeout" option for ClusterSet.forcePrimaryCluster()
and dba.rebootClusterFromCompleteOutage(), which defaults to shell
option "dba.gtidWaitTimeout", to wait for the applier to finish, making
sure that GTID_EXECUTED has the most up-to-date GTID set.

Included in this patch, ClusterSet.setPrimaryCluster() now also checks
for the applier to finish using the already existing "timeout" option.

Change-Id: I7fac18377e6dfad2999348774d04047817f1d2da