Adding support to pass python objects as parameters on the plugins.
The python value conversion was limited to the values representable by
the values supported in shcore::Value, which included:

- The primitive data types
- Arrays/Dictionaries
- Objects defined in Shell context (Session, Result, Etc)

Any value coming from python that did not fall into these categores was
translated to its string representation. This caused that objects
defined in python could not be passed as parameters to the plugin
functions.

This patch removes this restriction by wrapping python objects into
references when passed as parameters, rather than using the object
string representation.

Change-Id: If783f9566a2f717db73052d972b20981ed1efa00