BUG#34556560 consistency validation during dump too sensitive
The GTID-based consistency check which is used when the backup lock
cannot be acquired was too sensitive, it would report errors even if no
actual DDL changes were made. With this commit, the consistency check
analyzes which statements were executed during the dump, and fails only
if a DDL-related statement is detected.

A new option has been added to the dump utilities:
skipConsistencyChecks. It causes the dumper to skip the additional
consistency checks (scanning contents of the binlog to find potentially
problematic transactions) and assume that dump is consistent even though
gtid_executed/binlog position has changed.

Change-Id: I0fd95caa4e2fee35bf859eb6cbd4dd9d0315d2fe